moved methods to service
