+++ app/uk/gov/hmrc/incometaxsessiondata/services/SessionService.scala
@@ -16,7 +16,10 @@
 
 package uk.gov.hmrc.incometaxsessiondata.services
 
-import uk.gov.hmrc.incometaxsessiondata.models.{FullDuplicate, PartialDuplicate, NonDuplicate, Session, SessionData, SessionDataRequest, SessionDuplicationType}
+import play.api.Logging
+import play.api.mvc.Result
+import play.api.mvc.Results.{Conflict, Forbidden, InternalServerError, Ok}
+import uk.gov.hmrc.incometaxsessiondata.models.{FullDuplicate, NonDuplicate, PartialDuplicate, Session, SessionData, SessionDataRequest, SessionDuplicationType}
 import uk.gov.hmrc.incometaxsessiondata.repositories.SessionDataRepository
 
 import javax.inject.{Inject, Singleton}
@@ -26,7 +29,7 @@ import scala.util.{Failure, Success, Try}
 @Singleton
 class SessionService @Inject() (
   val repository: SessionDataRepository
-)(implicit ec: ExecutionContext) {
+)(implicit ec: ExecutionContext) extends Logging {
 
   def get(request: SessionDataRequest[_]): Future[Option[SessionData]] = {
     repository.get(request) map {
@@ -57,5 +60,63 @@ class SessionService @Inject() (
   }
 
   private case class IndexFields(sessionId: String, internalId: String)
+
+  def handleValidRequest(validRequest: Session): Future[Result] = {
+    getDuplicationStatus(validRequest) flatMap {
+      case result@FullDuplicate =>
+        logger.info(
+          s"[SessionController][handleValidRequest]" +
+            s" A session in the database matched the current session request, list of sessions: $result"
+        )
+        handleConflictScenario(validRequest)
+      case PartialDuplicate =>
+        logger.info(
+          s"[SessionController][handleValidRequest]" +
+            s" Another document matching mtditid: ${validRequest.mtditid} but different sessionId: ${validRequest.sessionId} and internalId ${validRequest.internalId}"
+        )
+        Future.successful(
+          Forbidden(
+            s"Another document matching mtditid: ${validRequest.mtditid} but different sessionId: ${validRequest.sessionId} and internalId ${validRequest.internalId}"
+          )
+        )
+      case NonDuplicate =>
+        logger.info(s"[SessionController][handleValidRequest]: No live sessions matching mtditid: ${validRequest.mtditid}")
+        handleOkScenario(validRequest)
+    }
+  }
+
+  private def handleConflictScenario(validRequest: Session): Future[Result] = {
+    set(validRequest) map {
+      case Right(true) =>
+        logger.info(
+          s"[SessionController][handleConflictScenario]: Successfully set session despite matching documenting existing in the database"
+        )
+        Conflict("Successfully set session despite matching documenting existing in the database")
+      case Right(false) =>
+        logger.info(s"[SessionController][handleConflictScenario]: Write operation was not acknowledged")
+        InternalServerError("Write operation was not acknowledged")
+      case Left(ex: Throwable) =>
+        logger.error(
+          s"[SessionController][handleConflictScenario]: Unknown exception < Message: ${ex.getMessage}, Cause: ${ex.getCause} >"
+        )
+        InternalServerError("Unknown exception")
+    }
+  }
+
+  private def handleOkScenario(validRequest: Session): Future[Result] = {
+    set(validRequest) map {
+      case Right(true) =>
+        logger.info(s"[SessionController][handleOkScenario]: Successfully set session")
+        Ok("Successfully set session")
+      case Right(false) =>
+        logger.info(s"[SessionController][handleOkScenario]: Write operation was not acknowledged")
+        InternalServerError("Write operation was not acknowledged")
+      case Left(ex: Throwable) =>
+        logger.error(
+          s"[SessionController][handleOkScenario]: Unknown exception < Message: ${ex.getMessage}, Cause: ${ex.getCause} >"
+        )
+        InternalServerError("Unknown exception")
+    }
+  }
 
 }
